requirements.txt
python-dotenv
telebot
gradio_client
wikipedia